Control left and right side of intake separately
Add new subsystems for left and right side of intake.  Need to do this
to allow left and right to spin (or not) independently.  A single
subsystem allows only one command at a time to be running.

To save some typing and try to organize code a bit, I left Intake
class as a subsystem but with no default command.  Left/Right sides of
intake inherit from Intake.  Not sure if this will work.  If it
doesn't, then need to move code from Intake.java into left/right
intake files for each side.