Get practice bot working with lastest real robot code.
Remove use of subystems not present on practice bot.
Remove button commands that control missing subystems.

Steps take to remove subystems and related code:
- comment out unused subsystem hw in RobotMap.java
- comment out subsystem var delcarations in Robot.java
  - This causes a bunch of compile errors to appear in Robot.java, DriveTrain.java, and commands that control those subsystems.
  - Comment out breaking/bad code in Robot.java and DriveTrain.java.
  - Leave commands alone.
  - Uncomment subsystem var declarations in Robot.java. This is remove errors in commands.

Goal is to minimize work removing code.  You could start commenting out and/or deleting code/files instead.  I chose to try to make minimal changes to disable subsystems not present.  None of this is tested so not sure it will work.